% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParams.R
\name{BDredux}
\alias{BDredux}
\title{Get BD diversification parameters from trees}
\usage{
BDredux(treeset)
}
\arguments{
\item{treeset}{Set of (probably empirical) phylogenies, list or multiPhylo-object.}
}
\value{
A dataframe with BD parameters.
}
\description{
Internal function used by \code{GetParams} to estimate BD parameters from a supplied tree or tree set.
}
\details{
The function uses \code{birthdeath} from the package \code{ape}.
}
