% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoskR.R
\docType{package}
\name{BoskR-package}
\alias{BoskR}
\alias{BoskR-package}
\title{BoskR - Assess Adequacy of Diversification Models Using Tree Shapes}
\description{
Given a phylogeny (or several) and a diversification model (or several), the package makes use of several tree shape metrics to infer summary statistics. The overlap of those metrics will be compared to the metrics from a set of trees simulated under the model in question and based on the same parameters as the initial tree. If the metrics indicate that the shapes of the simulated trees significantly differ from the initial one, the model should be deemed not adequate for this tree.
}
\details{
\packageIndices{BoskR}
}
\references{
Schwery O., O’Meara B.C. (2020) BoskR-Testing Adequacy of Diversification Models Using Tree Shape. bioRxiv.: 2020.12.21.423829
}
\seealso{
\code{\link{CombineTrees}}, \code{\link{GetMetricTreeSets}}, \code{\link{GetTreeMetrics}}, \code{\link{GetTreeParams}}, \code{\link{PvalMetrics}}, \code{\link{ScatterMetrics}}, \code{\link{TreeCorr}}, \code{\link{plotPvalMetricsCDF}}, \code{\link{plotPvalMetricsPDF}}
}
\author{
\strong{Maintainer}: Orlando Schwery \email{schwery.macroevo@pm.me} (\href{https://orcid.org/0000-0001-7373-7407}{ORCID})

}
\keyword{package}
