% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParams.R
\name{DDredux}
\alias{DDredux}
\title{Get DD diversification parameters from trees}
\usage{
DDredux(treeset)
}
\arguments{
\item{treeset}{Set of (probably empirical) phylogenies, list or multiPhylo-object.}
}
\value{
A dataframe with TimeD-BD parameters.
}
\description{
Internal function used by \code{GetParams} to estimate DD parameters from a supplied tree or tree set.
}
\details{
The function uses \code{dd_ML} from the package \code{DDD}.
}
