% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetricTrees.R
\name{DDtreeSim}
\alias{DDtreeSim}
\title{Get several DD trees out of DDD}
\usage{
DDtreeSim(numbsim, lambda, mu, K, age, ddmodel)
}
\arguments{
\item{numbsim}{Number of simulations per tree}

\item{lambda}{Speciation rate}

\item{mu}{Extinction rate}

\item{K}{Carrying capacity}

\item{age}{Crown age for simulated tree}

\item{ddmodel}{Model for dependence of diversification rates on K, see \code{dd_sim} description.}
}
\value{
A set of simulated trees
}
\description{
Internal function used by \code{GetMetricTrees} to simulate several trees under DD for each supplied tree or tree set.
}
\details{
The function uses \code{dd_sim} from the package \code{DDD}.
}
