% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptesttrees.R
\docType{data}
\name{emptesttrees}
\alias{emptesttrees}
\title{Set of Test Trees for BoskR}
\format{A list of phylogenies (of class phy)}
\source{
ADD THEM HERE
}
\usage{
data(epmtesttrees)
}
\description{
ADD DESCRIPTION HERE
}
\references{
ADD THEM HERE
}
\keyword{datasets}
