% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetricTrees.R
\name{GetMetricTreeSets}
\alias{GetMetricTreeSets}
\title{Simulate trees based on empirical estimations or set parameters}
\usage{
GetMetricTreeSets(empirical_start = FALSE, current_case = NULL,
  empParams = empParams, current_method, N = NULL, Numbsim1, Lambda,
  Mu, l = NULL, a = NULL, LambdaFun = NULL, MuFun = NULL,
  TreeAge = NULL, BiSSEpars = NULL, tree = NULL)
}
\arguments{
\item{empirical_start}{\code{TRUE} to use parameters estimated from empirical trees, \code{FALSE} to use user-specified ones}

\item{current_case}{Integer indicating which tree set tree belongs to (used for labeling purposes).}

\item{empParams}{Nested list object with tree parameters as inferred through \code{GetParams} from one or several empirical trees}

\item{current_method}{Method to be used for simulation, either \code{"BD", "TimeD-BD", "DD", "CD", "TraitD"} for birth-death, time-dependent birth-death, diversity dependent, clade dependent, or trait dependent diversification respectively.}

\item{N}{Number of taxa}

\item{Numbsim1}{Number of trees to simulate per each}

\item{Lambda}{Speciation rate}

\item{Mu}{Extinction rate}

\item{l}{Speciation rate}

\item{a}{Extinction fracion (Mu/Lambda)}

\item{LambdaFun}{Function for speciation rate}

\item{MuFun}{Function for extinction rate}

\item{TreeAge}{Stem age of tree}

\item{BiSSEpars}{Parameters from BiSSE}

\item{tree}{Phylogeny}
}
\value{
A list of trees of class multiPhylo
}
\description{
Uses \code{GetMetricTrees} to simulate trees under a given model based on either parameter estimates from empirical trees or pre-set parameters.
}
\details{
The function will simulate a number of trees based on either the parameters inferred from one or several empirical trees (given through \code{empParams} if \code{empirical_start=TRUE}), or user-specified parameters (if \code{empirical_start=FALSE})
}
