% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParams.R
\name{GetParams}
\alias{GetParams}
\title{Get diversification parameters from trees}
\usage{
GetParams(emptrees, current_method_est)
}
\arguments{
\item{emptrees}{Set of (probably empirical) phylogenies, list or multiPhylo-object.}

\item{current_method_est}{String specifying the method to be used to estimate the parameters. Can be \code{"BD", "TimeD-BD", "DD", "CD", "TraitD"} for birth-death, time-dependent birth-death, diversity dependent, clade dependent, or trait dependent diversification respectively.}
}
\value{
A nested list of parameter estimates for every tree in \code{emptrees}.
}
\description{
\code{GetParams} estimates parameters from a supplied tree or tree set, which can subsequently be used as input for tree simulations using \code{GetMetricTrees}.
}
\details{
The function uses either
}
