% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTreeMetrics.R
\name{GetTreeMetrics}
\alias{GetTreeMetrics}
\title{Get metrics describing tree shape}
\usage{
GetTreeMetrics(trees, empirical_start = FALSE)
}
\arguments{
\item{trees}{Tree or set of trees, list or multiPhylo-object}

\item{empirical_start}{\code{TRUE} if started out from empirical trees, \code{FALSE} if started from user-specified parameters}
}
\value{
A list with two elements: \code{metrics}: a matrix with the values for all tree metrics for each tree, and \code{spectra}: a list of raw values for the standard and normalised graph Laplacian spectra for each tree.
}
\description{
\code{GetTreeMetrics} calculates a number of metrics describing tree shape for a tree or a set of trees.
}
\details{
The function will calculate five 'traditional' tree metrics (Colless, Sackin, number of cherries, number of pitchforks, ladder sizes), as well as standard and normalised graph Laplacian spectra and the associated summary metrics (principal eigenvalue, asymmetry, peakedness, eigengap), as implemented in \code{RPANDA}.
}
