% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTreeParams.R
\name{GetTreeParams}
\alias{GetTreeParams}
\title{Get diversification parameters from trees}
\usage{
GetTreeParams(trees, current_method_est)
}
\arguments{
\item{trees}{Tree or set of trees, list or multiPhylo-object, or list of tree sets}

\item{current_method_est}{String specifying the method to be used to estimate the parameters. Can be \code{"BD", "TimeD-BD", "DD", "CD", "TraitD"} for birth-death, time-dependent birth-death, diversity dependent, clade dependent, or trait dependent diversification respectively.}
}
\value{
A nested list of parameter estimates for every tree in \code{trees}, or every tree in each tree set therein respectively.
}
\description{
\code{GetTreeParams} estimates parameters from a supplied tree or tree set, which can subsequently be used as input for tree simulations using \code{GetMetricTrees}.
}
\details{
The function wraps around the internal \code{GetParams}, and uses either
}
