% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPvalMetrics.R
\name{plotPvals}
\alias{plotPvals}
\title{Plot p-values for tree metrics}
\usage{
plotPvals(pmetrics, set, inloop = FALSE)
}
\arguments{
\item{pmetrics}{Object with ECDs and p-values of empirical and simulated tree shapes, output of \code{PvalMetrics} or formatted the same way.}

\item{set}{Numerical index for which of the sets of pairs of empirical and simulated metrics to be plotted.}

\item{inloop}{Logical indicating whether the function is called from within a loop (TRUE) or not (FALSE).}
}
\value{
An array of plots.
}
\description{
Internal function to create plots of p-values based on simulated and empirical distributions of tree metrics
}
