% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPvalMetricsPDF.R
\name{plotPvalsPDF}
\alias{plotPvalsPDF}
\title{Plot p-values for tree metrics}
\usage{
plotPvalsPDF(empMetrics, simMetrics, set, inloop = FALSE)
}
\arguments{
\item{empMetrics}{Metrics of empirical tree or set of trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{simMetrics}{Metrics of sets of simulated trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{set}{Numerical index for which of the sets of pairs of empirical and simulated metrics to be plotted.}

\item{inloop}{Logical indicating whether the function is called from within a loop (TRUE) or not (FALSE).}
}
\value{
An array of plots.
}
\description{
Internal function to create plots of p-values based on simulated and empirical distributions of tree metrics
}
