% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParams.R
\name{RateEstTraitDep}
\alias{RateEstTraitDep}
\title{Get TraitD diversification parameters from trees}
\usage{
RateEstTraitDep(treeset)
}
\arguments{
\item{treeset}{Set of (probably empirical) phylogenies, list or multiPhylo-object.}
}
\value{
A dataframe with TimeD-BD parameters.
}
\description{
Internal function used by \code{GetParams} to estimate TraitD parameters from a supplied tree or tree set.
}
\details{
The function uses \code{make.bisse} and \code{find.mle} from the package \code{diversitree}.
}
