% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParams.R
\name{TimeDBDredux}
\alias{TimeDBDredux}
\title{Get TimeD-BD diversification parameters from trees}
\usage{
TimeDBDredux(treeset)
}
\arguments{
\item{treeset}{Set of (probably empirical) phylogenies, list or multiPhylo-object.}
}
\value{
A dataframe with TimeD-BD parameters.
}
\description{
Internal function used by \code{GetParams} to estimate TimeD-BD parameters from a supplied tree or tree set.
}
\details{
The function uses \code{fitSPVAR} from the package \code{laser}.
}
