\name{BoskR-package}
\alias{BoskR-package}
\alias{BoskR}
\docType{package}
\title{
\packageTitle{BoskR}
}
\description{
\packageDescription{BoskR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BoskR}
\packageIndices{BoskR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BoskR}

Maintainer: \packageMaintainer{BoskR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
