\name{BoskR-package}
\alias{BoskR-package}
\alias{BoskR}
\docType{package}
\title{
\packageTitle{BoskR}
}
\description{
\packageDescription{BoskR}
}
\details{
\packagePackage{BoskR},

\packageType{BoskR},

\packageVersion{BoskR},

\packageDate{BoskR},

\packageLicense{BoskR},

\packageIndices{BoskR}
}
\author{
\packageAuthors@R{BoskR}
}
\references{
In prep.
}
\keyword{ package }
\seealso{
\code{\link{CombineTrees}}, \code{\link{GetMetricTreeSets}}, \code{\link{GetTreeMetrics}}, \code{\link{GetTreeParams}}, \code{\link{PvalMetrics}}, \code{\link{ScatterMetrics}}, \code{\link{TreeCorr}}, \code{\link{plotPvalMetricsCDF}}, \code{\link{plotPvalMetricsPDF}}

\code{\link[RPANDA:RPANDA-package]{RPANDA}}
}
