% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTreeMetrics.R
\name{GetTreeMetrics}
\alias{GetTreeMetrics}
\title{Get metrics describing tree shape}
\usage{
GetTreeMetrics(trees, empirical_start = FALSE, source = c("emp"))
}
\arguments{
\item{trees}{Tree or set of trees, list or multiPhylo-object, or list of tree sets}

\item{empirical_start}{\code{TRUE} if started out from empirical trees, \code{FALSE} if started from user-specified parameters}

\item{source}{Character string to indicate whether \code{trees} is the empirical or simulated input trees (i.e. most likely a tree or set of trees), in which case it should be "emp", or the sets of simulated trees based on those input trees (i.e. a list of tree sets), in which case it should be "sim".}
}
\value{
A list with two elements: \code{metrics}: a matrix with the values for all tree metrics for each tree, and \code{spectra}: a list of raw values for the standard and normalised graph Laplacian spectra for each tree. If \code{source="sim"}, it will be one such two-element list for each tree set provided in a nested list.
}
\description{
\code{GetTreeMetrics} calculates a number of metrics describing tree shape for a tree or a set of trees.
}
\details{
The function wraps around the internal 'GetMetrics', which will calculate five 'traditional' tree metrics (Colless, Sackin, number of cherries, number of pitchforks, ladder sizes), as well as standard and normalised graph Laplacian spectra and the associated summary metrics (principal eigenvalue, asymmetry, peakedness, eigengap), as implemented in \code{RPANDA}.
}
