% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPvalMetricsCDF.R
\name{plotPvalMetricsCDF}
\alias{plotPvalMetricsCDF}
\title{Plot p-values on CDF for sets of tree metrics}
\usage{
plotPvalMetricsCDF(pmetrics, set = NULL)
}
\arguments{
\item{pmetrics}{Object with ECDs and p-values of empirical and simulated tree shapes, output of \code{PvalMetrics} or formatted the same way.}

\item{set}{Numerical index for which of the sets of pairs of empirical and simulated metrics to be plotted; default NULL will plot all sets.}
}
\value{
An array of plots.
}
\description{
Creates plots of p-values on their corresponding cumulative distribution function, based on sets of simulated and empirical distributions of tree metrics,
}
