% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PvalMetrics.R
\name{PvalMetrics}
\alias{PvalMetrics}
\title{Get p-values for tree metrics}
\usage{
PvalMetrics(empMetrics, simMetrics, current_case, methodnr)
}
\arguments{
\item{empMetrics}{Metrics of empirical tree or set of trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{simMetrics}{Metrics of sets of simulated trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{current_case}{Indicator whether empMetrics is based of simulated initial trees, in which case the value should be "simed"; mainly important for data format reasons.}

\item{methodnr}{Integral specifying which method is used: 1: BD, 2: TimeD-BD, 3: DD.}
}
\value{
A list with two entries: \code{ECDs} is a list of Empirical Cumulative Distributions; \code{pValues} is a matrix with p-values for the targeted metrics
}
\description{
Estimates p-values based on simulated and empirical distributions of tree metrics
}
\details{
The function uses an Empirical Cumulative Distribution function to determine the area under the curve of the metric values of the simulated trees, to get to a p-value for the position of the metrics of the empirical tree on that distribution.
}
