% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterMetrics.R
\name{ScatterMetrics}
\alias{ScatterMetrics}
\title{3D Metrics Scatterplot}
\usage{
ScatterMetrics(
  empMetrics,
  simMetrics,
  pair = 1,
  skim = FALSE,
  combine = FALSE,
  colours = c("black", "red"),
  transparencyEmp = 0.8,
  transparencySim = 0.2,
  pch = 16,
  cex.symbols = 1.5,
  main = paste("Empirical vs. Simulated Metrics Set", pair, sep = " "),
  angle = -230
)
}
\arguments{
\item{empMetrics}{Metrics of empirical tree or set of trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{simMetrics}{Metrics of sets of simulated trees; output of \code{GetTreeMetrics} or formatted the same way.}

\item{pair}{Numerical index for which of the sets of pairs of empirical and simulated metrics to be plotted. Value is ignored if \code{skim} or \code{combine} are \code{TRUE}.}

\item{skim}{Logical, creates interactive plot of all pairs of empirical trees and their simulations if \code{TRUE}; one can advance through the plots by hitting enter.}

\item{combine}{Logical, combines all empirical and simulated trees into one plot if \code{TRUE}.}

\item{colours}{Vector of length two, indicating the desired colours for empirical trees and simulated treees, in that order (defaults are "black" and "red", respectively).}

\item{transparencyEmp}{Value determining the transparency of the empirical tree plot points (0: completely transparent, 1: completely opaque; corresponding to \code{alpha} from package \code{scales}).}

\item{transparencySim}{Value determining the transparency of the simulated tree plot points (0: completely transparent, 1: completely opaque; corresponding to \code{alpha} from package \code{scales}).}

\item{pch}{Shape of plot symbols; default 16.}

\item{cex.symbols}{Size of plot symbols; default 1.5.}

\item{main}{String for plot title; default "Empirical vs. Simulated Metrics Set", followed by pair number plotted, or "Combined".}

\item{angle}{Rotation of the plot, determined by angle between x and y axis (corresponding to \code{scatterplot3d}); default -230.}
}
\value{
3D scatterplot of trees in metric space, or a series of such plots to skip through.
}
\description{
Plots empirical trees and their simulations in tree metric space using a 3D scatterplot.
}
\details{
The function uses the internals \code{ScatterMetricsPair} and \code{ScatterMetricsCombo} and plots the empirical input-trees and their corresponding simulations in the metric space (asymmetry x peakedness x principal Eigenvalue) as a 3D scatterplot. It allows to either plot them all combined, or pairwise. The latter meaning each empirical tree is plotted with its corresponding simulations only, either one at a time or all together interactively (one advances through the plots by pressing enter). The basic function used is \code{scatterplot3d}, from the package with the same name.
}
